+++ HISTORY.md
@@ -14,6 +14,7 @@ v0.7 - 2014/06/18
 * New channel reports: comments, likes, dislikes, shares and impressions
 * Allow both normal and partnered channels to retrieve reports about views, comments, likes, dislikes, shares
 * Make reports available also on Video (not just Channel)
+* New account.upload_video to upload a video (either local or remote).
 
 v0.6 - 2014/06/05
 -----------------